chatgpt-bot/requirements.txt
openai == 1.3.7
langchain
line-bot-sdk >= 3.5.1
typing_extensions
